+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 3.4.0
 -----
 
+ * Added guard `is_valid()` method support.
  * Added support for `Event::getWorkflowName()` for "announce" events.
  * Added `workflow.completed` events which are fired after a transition is completed.